LPS-184766 Activator: Move package, no logic change
