+++ modules/apps/static/portal-osgi-web/portal-osgi-web-http-servlet/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Activator: com.liferay.portal.osgi.web.http.servlet.internal.HttpServletBundleActivator
+Bundle-Activator: com.liferay.portal.osgi.web.http.servlet.internal.activator.HttpServletBundleActivator
 Bundle-Name: Liferay Portal OSGi Web Http Servlet
 Bundle-SymbolicName: com.liferay.portal.osgi.web.http.servlet
 Bundle-Version: 1.0.0
