+++ modules/apps/static/portal-osgi-web/portal-osgi-web-http-servlet/src/main/java/com/liferay/portal/osgi/web/http/servlet/internal/activator/HttpServletBundleActivator.java
@@ -1,17 +1,18 @@
-/*******************************************************************************
- * Copyright (c) 2005, 2015 Cognos Incorporated, IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
- * Contributors:
- *     Cognos Incorporated - initial API and implementation
- *     IBM Corporation - bug fixes and enhancements
- *     Raymond Augé <raymond.auge@liferay.com> - Bug 436698
- *******************************************************************************/
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
 
-package com.liferay.portal.osgi.web.http.servlet.internal;
+package com.liferay.portal.osgi.web.http.servlet.internal.activator;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.log.Log;
@@ -19,6 +20,8 @@ import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.util.HashMapDictionary;
 import com.liferay.portal.osgi.web.http.servlet.ExtendedHttpService;
 import com.liferay.portal.osgi.web.http.servlet.HttpServiceServlet;
+import com.liferay.portal.osgi.web.http.servlet.internal.HttpServiceFactory;
+import com.liferay.portal.osgi.web.http.servlet.internal.HttpServiceRuntimeImpl;
 import com.liferay.portal.osgi.web.http.servlet.internal.servlet.ProxyServlet;
 import com.liferay.portal.osgi.web.http.servlet.internal.util.HttpTuple;
 import com.liferay.portal.osgi.web.http.servlet.internal.util.UMDictionaryMap;
@@ -46,9 +49,7 @@ import org.osgi.util.tracker.ServiceTracker;
 import org.osgi.util.tracker.ServiceTrackerCustomizer;
 
 /**
- * @author Cognos Incorporated
- * @author IBM Corporation
- * @author Raymond Augé
+ * @author Dante Wang
  */
 public class HttpServletBundleActivator implements BundleActivator {
